+++ engine/runner/src/main/scala/org/enso/runner/Main.scala
@@ -7,7 +7,7 @@ import com.typesafe.scalalogging.Logger
 import org.apache.commons.cli.{Option => CliOption, _}
 import org.enso.editions.DefaultEdition
 import org.enso.languageserver.boot
-import org.enso.languageserver.boot.LanguageServerConfig
+import org.enso.languageserver.boot.{LanguageServerConfig, ProfilingConfig}
 import org.enso.libraryupload.LibraryUploader.UploadFailedError
 import org.enso.loggingservice.LogLevel
 import org.enso.pkg.{Contact, PackageManager, Template}
@@ -16,14 +16,14 @@ import org.enso.version.VersionDescription
 import org.graalvm.polyglot.PolyglotException
 
 import java.io.File
-import java.nio.file.Path
-import java.util.UUID
+import java.nio.file.{Path, Paths}
+import java.util.{Collections, UUID}
 
 import scala.Console.err
+import scala.concurrent.duration._
 import scala.jdk.CollectionConverters._
-import scala.util.{Failure, Success, Try}
 import scala.util.control.NonFatal
-import java.util.Collections
+import scala.util.{Failure, Success, Try}
 
 /** The main CLI entry point class. */
 object Main {
@@ -40,7 +40,10 @@ object Main {
   private val DOCS_OPTION                    = "docs"
   private val PREINSTALL_OPTION              = "preinstall-dependencies"
   private val LANGUAGE_SERVER_OPTION         = "server"
-  private val LANGUAGE_SERVER_PROFILING      = "server-profiling"
+  private val LANGUAGE_SERVER_PROFILING_PATH = "server-profiling-path"
+  private val LANGUAGE_SERVER_PROFILING_TIME = "server-profiling-time"
+  private val LANGUAGE_SERVER_PROFILING_EVENTS_LOG_PATH =
+    "server-profiling-events-log-path"
   private val DAEMONIZE_OPTION               = "daemon"
   private val INTERFACE_OPTION               = "interface"
   private val RPC_PORT_OPTION                = "rpc-port"
@@ -153,11 +156,26 @@ object Main {
       .longOpt(LANGUAGE_SERVER_OPTION)
       .desc("Runs Language Server")
       .build()
-    val lsProfilingOption = CliOption.builder
-      .longOpt(LANGUAGE_SERVER_PROFILING)
-      .desc(
-        "Enables the Language Server profiling. The output is written to system temp directory."
-      )
+    val lsProfilingPathOption = CliOption.builder
+      .hasArg(true)
+      .numberOfArgs(1)
+      .argName("file")
+      .longOpt(LANGUAGE_SERVER_PROFILING_PATH)
+      .desc("The path to the Language Server profiling file.")
+      .build()
+    val lsProfilingTimeOption = CliOption.builder
+      .hasArg(true)
+      .numberOfArgs(1)
+      .argName("seconds")
+      .longOpt(LANGUAGE_SERVER_PROFILING_TIME)
+      .desc("The duration in seconds limiting the profiling time.")
+      .build()
+    val lsProfilingEventsLogPathOption = CliOption.builder
+      .hasArg(true)
+      .numberOfArgs(1)
+      .argName("file")
+      .longOpt(LANGUAGE_SERVER_PROFILING_EVENTS_LOG_PATH)
+      .desc("The path to the runtime events log file.")
       .build()
     val deamonizeOption = CliOption.builder
       .longOpt(DAEMONIZE_OPTION)
@@ -335,7 +353,9 @@ object Main {
       .addOption(newProjectAuthorNameOpt)
       .addOption(newProjectAuthorEmailOpt)
       .addOption(lsOption)
-      .addOption(lsProfilingOption)
+      .addOption(lsProfilingPathOption)
+      .addOption(lsProfilingTimeOption)
+      .addOption(lsProfilingEventsLogPathOption)
       .addOption(deamonizeOption)
       .addOption(interfaceOption)
       .addOption(rpcPortOption)
@@ -823,14 +843,29 @@ object Main {
       dataPort <- Either
         .catchNonFatal(dataPortStr.toInt)
         .leftMap(_ => "Port must be integer")
-      profilingEnabled = line.hasOption(LANGUAGE_SERVER_PROFILING)
+      profilingPathStr =
+        Option(line.getOptionValue(LANGUAGE_SERVER_PROFILING_PATH))
+      profilingPath <- Either
+        .catchNonFatal(profilingPathStr.map(Paths.get(_)))
+        .leftMap(_ => "Profiling path is invalid")
+      profilingTimeStr = Option(
+        line.getOptionValue(LANGUAGE_SERVER_PROFILING_TIME)
+      )
+      profilingTime <- Either
+        .catchNonFatal(profilingTimeStr.map(_.toInt.seconds))
+        .leftMap(_ => "Profiling time should be an integer")
+      profilingEventsLogPathStr =
+        Option(line.getOptionValue(LANGUAGE_SERVER_PROFILING_EVENTS_LOG_PATH))
+      profilingEventsLogPath <- Either
+        .catchNonFatal(profilingEventsLogPathStr.map(Paths.get(_)))
+        .leftMap(_ => "Profiling events log path is invalid")
     } yield boot.LanguageServerConfig(
       interface,
       rpcPort,
       dataPort,
       rootId,
       rootPath,
-      profilingEnabled
+      ProfilingConfig(profilingEventsLogPath, profilingPath, profilingTime)
     )
 
   /** Prints the version of the Enso executable.