+++ lib/scala/profiling-utils/src/main/scala/org/enso/profiling/FileSampler.scala
@@ -3,10 +3,9 @@ package org.enso.profiling
 import org.netbeans.modules.sampler.Sampler
 
 import java.io.{DataOutputStream, File, FileOutputStream}
-import java.nio.file.Files
 import java.util.concurrent.{CompletableFuture, Executor, Executors}
 
-import scala.concurrent.duration.Duration
+import scala.concurrent.duration.FiniteDuration
 import scala.util.Using
 
 /** Gathers application performance statistics that can be visualised in Java
@@ -14,15 +13,10 @@ import scala.util.Using
   *
   * @param output the output stream to write the collected statistics to
   */
-final class TempFileSampler(output: File) extends MethodsSampler {
+final class FileSampler(output: File) extends MethodsSampler {
 
   private val sampler: Sampler         = Sampler.createSampler(getClass.getSimpleName)
   private var samplingStarted: Boolean = false
-
-  def getSiblingFile(ext: String): File = {
-    val newName = output.getName.replace(".npss", ext)
-    new File(output.getParent, newName)
-  }
 
   /** @inheritdoc */
   def start(): Unit =
@@ -45,7 +39,7 @@ final class TempFileSampler(output: File) extends MethodsSampler {
     }
 
   /** @inheritdoc */
-  def stop(delay: Duration)(implicit ec: Executor): Unit =
+  def stop(delay: FiniteDuration)(implicit ec: Executor): Unit =
     this.synchronized {
       val executor = Executors.newSingleThreadScheduledExecutor()
 
@@ -58,22 +52,4 @@ final class TempFileSampler(output: File) extends MethodsSampler {
         )
         .whenComplete((_, _) => executor.shutdown())
     }
-
-  /** @return `true` if the sampling is started. */
-  def isSamplingStarted: Boolean =
-    this.samplingStarted
-}
-object TempFileSampler {
-
-  /** Create an instance of [[MethodsSampler]] that writes the data to the
-    * temporary `.npss` file with the provided prefix.
-    *
-    * @param prefix the prefix of the temp file.
-    * @return the [[MethodsSampler]] instance
-    */
-  def apply(prefix: String): TempFileSampler = {
-    val path = Files.createTempFile(s"$prefix-", ".npss")
-    Files.deleteIfExists(path)
-    new TempFileSampler(path.toFile)
-  }
 }