+++ std-bits/table/src/main/java/org/enso/table/read/DelimitedReader.java
@@ -8,37 +8,27 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import org.enso.table.data.column.builder.string.StorageBuilder;
 import org.enso.table.data.column.builder.string.StringStorageBuilder;
 import org.enso.table.data.column.storage.Storage;
+import org.enso.table.data.column.storage.StringStorage;
 import org.enso.table.data.index.DefaultIndex;
 import org.enso.table.data.table.Column;
 import org.enso.table.data.table.Table;
+import org.enso.table.parsing.DatatypeParser;
+import org.enso.table.parsing.TypeInferringParser;
 import org.enso.table.parsing.problems.AdditionalInvalidRows;
 import org.enso.table.parsing.problems.InvalidRow;
 import org.enso.table.parsing.problems.MismatchedQuote;
+import org.enso.table.parsing.problems.NoOpProblemAggregator;
 import org.enso.table.parsing.problems.ParsingProblem;
 import org.enso.table.util.NameDeduplicator;
 
 /** A helper for reading delimited (CSV-like) files. */
 public class DelimitedReader {
 
-  /** Specifies how to set the headers for the returned table. */
-  public enum HeaderBehavior {
-    /** Tries to infer if the headers are present in the file. */
-    INFER,
-
-    /** Uses the first row in the file as headers. Duplicate names will be appended suffixes. */
-    USE_FIRST_ROW_AS_HEADERS,
-
-    /**
-     * Treats the first row as data and generates header names starting with {@code COLUMN_NAME}.
-     */
-    GENERATE_HEADERS
-  }
-
   private static final String COLUMN_NAME = "Column";
-
+  private static final char noQuoteCharacter = '\0';
+  private static final long invalidRowsLimit = 10;
   private final char delimiter;
   private final char quoteCharacter;
   private final char quoteEscapeCharacter;
@@ -48,10 +38,16 @@ public class DelimitedReader {
   private final int maxColumns;
   private final List<ParsingProblem> warnings = new ArrayList<>();
   private final CsvParser parser;
+  private final DatatypeParser valueParser;
+  private final TypeInferringParser cellTypeGuesser;
   private final boolean keepInvalidRows;
   private final boolean warningsAsErrors;
-
-  private static final char noQuoteCharacter = '\0';
+  private final NoOpProblemAggregator noOpProblemAggregator = new NoOpProblemAggregator();
+  private long invalidRowsCount = 0;
+  private long targetTableIndex = 0;
+  /** The line number of the start of the current row in the input file. */
+  private long currentLine = 0;
+  private StringStorageBuilder[] builders = null;
 
   /**
    * Creates a new reader.
@@ -70,6 +66,10 @@ public class DelimitedReader {
    * @param skipRows specifies how many rows from the input to skip
    * @param rowLimit specifies how many rows to read (does not include the header row)
    * @param maxColumns specifies how many columns can be expected at most
+   * @param valueParser an optional parser that is applied to each column to convert it to more
+   *     specific datatype
+   * @param cellTypeGuesser a helper used to guess cell types, used for the purpose of inferring the
+   *     headers, it must not be null if {@code headerBehavior} is set to {@code INFER}.
    * @param keepInvalidRows specifies whether to keep rows that had an unexpected number of columns
    * @param warningsAsErrors specifies if the first warning should be immediately raised as an error
    *     (used as a fast-path for the error-reporting mode to avoid computing a value that is going
@@ -84,6 +84,8 @@ public class DelimitedReader {
       long skipRows,
       long rowLimit,
       int maxColumns,
+      DatatypeParser valueParser,
+      TypeInferringParser cellTypeGuesser,
       boolean keepInvalidRows,
       boolean warningsAsErrors) {
     if (delimiter.isEmpty()) {
@@ -136,6 +138,8 @@ public class DelimitedReader {
     this.keepInvalidRows = keepInvalidRows;
     this.warningsAsErrors = warningsAsErrors;
 
+    this.valueParser = valueParser;
+    this.cellTypeGuesser = cellTypeGuesser;
     parser = setupCsvParser(input);
   }
 
@@ -158,54 +162,15 @@ public class DelimitedReader {
     return parser;
   }
 
-  /** Parses a cell, removing surrounding quotes (if applicable). */
-  private String parseCell(String cell) {
-    if (cell == null) return null;
-
-    if (cell.isEmpty()) return cell;
-    if (cell.charAt(0) == quoteCharacter) {
-      return stripQuotes(cell);
-    }
-
-    return cell;
-  }
-
   /** Parses a header cell, removing surrounding quotes (if applicable). */
   private String parseHeader(String cell) {
     if (cell == null) return COLUMN_NAME;
-
-    if (cell.isEmpty()) return cell;
-    if (cell.charAt(0) == quoteCharacter) {
-      return stripQuotes(cell);
-    }
-
-    return cell;
-  }
-
-  /**
-   * If the first character of a string is a quote, will remove the surrounding quotes.
-   *
-   * <p>If the first character of a string is a quote but the last one is not, mismatched quote
-   * problem is reported.
-   */
-  private String stripQuotes(String cell) {
-    assert cell.charAt(0) == quoteCharacter;
-
-    if (cell.length() < 2 || cell.charAt(cell.length() - 1) != quoteCharacter) {
-      reportMismatchedQuote();
-      return cell.substring(1);
-    } else {
-      // Strip quotes.
-      return cell.substring(1, cell.length() - 1);
-    }
+    return QuoteHelper.stripQuotes(quoteCharacter, this::reportMismatchedQuote, cell);
   }
 
   private void reportMismatchedQuote() {
     reportProblem(new MismatchedQuote());
   }
-
-  private long invalidRowsCount = 0;
-  private static final long invalidRowsLimit = 10;
 
   private void reportInvalidRow(long source_row, Long table_index, String[] row) {
     if (invalidRowsCount < invalidRowsLimit) {
@@ -233,29 +198,89 @@ public class DelimitedReader {
     }
   }
 
-  private long target_table_index = 0;
-
-  /** The line number of the start of the current row in the input file. */
-  private long current_line = 0;
-
   /**
    * Reads the next row and updates the current line accordingly.
    *
    * <p>Will return {@code null} if no more rows are available.
    */
-  private String[] nextRow() {
-    current_line = parser.getContext().currentLine() + 1;
+  private String[] readNextRow() {
+    currentLine = parser.getContext().currentLine() + 1;
     return parser.parseNext();
+  }
+
+  private void appendRow(String[] row) {
+    assert builders != null;
+    assert canFitMoreRows();
+
+    if (row.length != builders.length) {
+      reportInvalidRow(currentLine, keepInvalidRows ? targetTableIndex : null, row);
+
+      if (keepInvalidRows) {
+        for (int i = 0; i < builders.length && i < row.length; i++) {
+          builders[i] = builders[i].parseAndAppend(row[i]);
+        }
+
+        // If the current row had less columns than expected, nulls are inserted for the missing
+        // values.
+        // If it had more columns, the excess columns are discarded.
+        for (int i = row.length; i < builders.length; i++) {
+          builders[i] = builders[i].parseAndAppend(null);
+        }
+
+        targetTableIndex++;
+      }
+    } else {
+      for (int i = 0; i < builders.length; i++) {
+        builders[i] = builders[i].parseAndAppend(row[i]);
+      }
+
+      targetTableIndex++;
+    }
+  }
+
+  private boolean canFitMoreRows() {
+    return rowLimit < 0 || targetTableIndex < rowLimit;
+  }
+
+  private void appendRowIfLimitPermits(String[] row) {
+    if (canFitMoreRows()) {
+      appendRow(row);
+    }
+  }
+
+  private List<String> headersFromRow(String[] row) {
+    List<String> preprocessedHeaders =
+        Arrays.stream(row).map(this::parseHeader).collect(Collectors.toList());
+    return NameDeduplicator.deduplicate(preprocessedHeaders, "_");
+  }
+
+  private List<String> generateDefaultHeaders(int columnCount) {
+    ArrayList<String> headerNames = new ArrayList<>(columnCount);
+    for (int i = 0; i < columnCount; ++i) {
+      headerNames.add(COLUMN_NAME + "_" + (i + 1));
+    }
+    return headerNames;
+  }
+
+  /**
+   * Checks if the given cell contains just plain text that is not null and is not convertible to
+   * any more specific type according to the {@code cellTypeGuesser}. This is used for checking the
+   * types when inferring the headers.
+   */
+  private boolean isPlainText(String cell) {
+    if (cell == null) return false;
+    Object parsed = cellTypeGuesser.parseSingleValue(cell, noOpProblemAggregator);
+    return parsed instanceof String;
   }
 
   /** Reads the input stream and returns a Table. */
   public Table read() {
     List<String> headerNames;
-    String[] currentRow = nextRow();
+    String[] currentRow = readNextRow();
 
     // Skip the first N rows.
     for (long i = 0; currentRow != null && i < skipRows; ++i) {
-      currentRow = nextRow();
+      currentRow = readNextRow();
     }
 
     // If there are no rows to even infer the headers, we return an empty table.
@@ -263,75 +288,88 @@ public class DelimitedReader {
       return new Table(new Column[0]);
     }
 
+    int expectedColumnCount = currentRow.length;
+    initBuilders(expectedColumnCount);
+
+    assert currentRow != null;
     switch (headerBehavior) {
-      case INFER:
-        throw new IllegalStateException("Inferring headers is not yet implemented");
-      case USE_FIRST_ROW_AS_HEADERS:
-        List<String> preprocessedHeaders =
-            Arrays.stream(currentRow).map(this::parseHeader).collect(Collectors.toList());
-        headerNames = NameDeduplicator.deduplicate(preprocessedHeaders, "_");
-        // We have 'used up' the first row, so we load a next one.
-        currentRow = nextRow();
-        break;
-      case GENERATE_HEADERS:
-        headerNames = new ArrayList<>(currentRow.length);
-        for (int i = 0; i < currentRow.length; ++i) {
-          headerNames.add(COLUMN_NAME + "_" + (i + 1));
+      case INFER -> {
+        String[] firstRow = currentRow;
+        String[] secondRow = readNextRow();
+        if (secondRow == null) {
+          // If there is only one row in the file, we generate the headers and stop further processing (as nothing more to process).
+          headerNames = generateDefaultHeaders(expectedColumnCount);
+          appendRowIfLimitPermits(firstRow);
+          currentRow = null;
+        } else {
+          assert cellTypeGuesser != null;
+          boolean firstAllText = Arrays.stream(firstRow).allMatch(this::isPlainText);
+          boolean secondAllText = Arrays.stream(secondRow).allMatch(this ::isPlainText);
+          boolean useFirstRowAsHeader = firstAllText && !secondAllText;
+          if (useFirstRowAsHeader) {
+            headerNames = headersFromRow(firstRow);
+            appendRowIfLimitPermits(secondRow);
+          } else {
+            headerNames = generateDefaultHeaders(expectedColumnCount);
+            appendRowIfLimitPermits(firstRow);
+            appendRowIfLimitPermits(secondRow);
+          }
+
+          currentRow = readNextRow();
         }
-        break;
-      default:
-        throw new IllegalStateException("Impossible branch.");
+      }
+      case USE_FIRST_ROW_AS_HEADERS -> {
+        headerNames = headersFromRow(currentRow);
+        // We have 'used up' the first row, so we load a next one.
+        currentRow = readNextRow();
+      }
+      case GENERATE_HEADERS -> {
+        headerNames = generateDefaultHeaders(expectedColumnCount);
+      }
+      default -> throw new IllegalStateException("Impossible branch.");
     }
 
-    StorageBuilder[] builders = initBuilders(headerNames.size());
-
-    while (currentRow != null && (rowLimit < 0 || target_table_index < rowLimit)) {
-      if (currentRow.length != builders.length) {
-        reportInvalidRow(current_line, keepInvalidRows ? target_table_index : null, currentRow);
-
-        if (keepInvalidRows) {
-          for (int i = 0; i < builders.length && i < currentRow.length; i++) {
-            String item = parseCell(currentRow[i]);
-            builders[i] = builders[i].parseAndAppend(item);
-          }
-
-          // If the current row had less columns than expected, nulls are inserted for the missing
-          // values.
-          // If it had more columns, the excess columns are discarded.
-          for (int i = currentRow.length; i < builders.length; i++) {
-            builders[i] = builders[i].parseAndAppend(null);
-          }
-
-          target_table_index++;
-        }
-      } else {
-        for (int i = 0; i < builders.length; i++) {
-
-          String item = parseCell(currentRow[i]);
-          builders[i] = builders[i].parseAndAppend(item);
-        }
-
-        target_table_index++;
-      }
-
-      currentRow = nextRow();
+    while (currentRow != null && canFitMoreRows()) {
+      appendRow(currentRow);
+      currentRow = readNextRow();
     }
 
     parser.stopParsing();
 
     Column[] columns = new Column[builders.length];
     for (int i = 0; i < builders.length; i++) {
-      Storage col = builders[i].seal();
-      columns[i] = new Column(headerNames.get(i), new DefaultIndex(col.size()), col);
+      String columnName = headerNames.get(i);
+      StringStorage col = builders[i].seal();
+
+      WithProblems<Storage> parseResult = valueParser.parseColumn(columnName, col);
+      for (var problem : parseResult.problems()) {
+        reportProblem(problem);
+      }
+      Storage storage = parseResult.value();
+
+      columns[i] = new Column(columnName, new DefaultIndex(storage.size()), storage);
     }
     return new Table(columns);
   }
 
-  private StorageBuilder[] initBuilders(int count) {
-    StorageBuilder[] res = new StorageBuilder[count];
+  private void initBuilders(int count) {
+    builders = new StringStorageBuilder[count];
     for (int i = 0; i < count; i++) {
-      res[i] = new StringStorageBuilder();
+      builders[i] = new StringStorageBuilder();
     }
-    return res;
+  }
+
+  /** Specifies how to set the headers for the returned table. */
+  public enum HeaderBehavior {
+    /** Tries to infer if the headers are present in the file. */
+    INFER,
+
+    /** Uses the first row in the file as headers. Duplicate names will be appended suffixes. */
+    USE_FIRST_ROW_AS_HEADERS,
+
+    /**
+     * Treats the first row as data and generates header names starting with {@code COLUMN_NAME}.
+     */
+    GENERATE_HEADERS
   }
 }